Make sure cache lock is being released
The node execution might case raise an exception
in that case the lock needs to be released since
no code is executing the locked task.

This bug was in particular annoing when an View
task raised a PermissionDenied exception.
Once you logged in the lock was still in place
and you faced a FlockLockFailed exception.